moved the modal-specific views and view models to their own modals sub-package
