+++ app/src/main/java/com/mikewarren/speakify/viewsAndViewModels/pages/importantApps/ConfigurableAppListItemViewModel.kt
@@ -3,7 +3,7 @@ package com.mikewarren.speakify.viewsAndViewModels.pages.importantApps
 import com.mikewarren.speakify.data.AppSettingsModel
 import com.mikewarren.speakify.data.SettingsRepository
 import com.mikewarren.speakify.data.UserAppModel
-import kotlinx.coroutines.flow.last
+import com.mikewarren.speakify.viewsAndViewModels.pages.importantApps.modals.AppSettingsViewModel
 
 class ConfigurableAppListItemViewModel(
     override val model: UserAppModel,
