+++ app/src/main/java/com/mikewarren/speakify/viewsAndViewModels/pages/importantApps/modals/DeleteConfirmationDialog.kt
@@ -1,4 +1,4 @@
-package com.mikewarren.speakify.viewsAndViewModels.pages.importantApps
+package com.mikewarren.speakify.viewsAndViewModels.pages.importantApps.modals
 
 import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Button
